refactor(viewers): split describe/log update files under 800-line cap
Extract diff handlers from update_describe.go into update_diff.go and
shared visual-copy helpers into update_visual_copy.go. Split the log
viewer's normal/visual/search handlers out of update_logs.go into
focused per-mode files, leaving only the top-level dispatchers behind.
